+++ sniffer/exceptions/network.py
@@ -6,3 +6,8 @@ class UnsupportedVersionException(Exception):
 class SLLUnsupportedError(Exception):
     """SLL is not supported."""
     pass
+
+
+class UninterestingPacketException(Exception):
+    """We only care about HTTP packets, not other things."""
+    pass