+++ sniffer/network/analyzer.py
@@ -1,11 +1,16 @@
-from .layers import Layer2, Layer3, Layer4
+from .layers import Layer2, Layer3, Layer4, Layer7
+from ..exceptions.network import UninterestingPacketException
 
 
 class PacketAnalyzer:
     def __init__(self, packet_bytes: bytes):
+        if b'HTTP/' not in packet_bytes:
+            raise UninterestingPacketException
+
         self.layer2 = Layer2(packet_bytes)
         self.layer3 = Layer3(packet_bytes)
         self.layer4 = Layer4(packet_bytes)
+        self.layer7 = Layer7(packet_bytes)
 
     def get_source_mac(self):
         return self.layer2.source_mac.decode('ascii')
@@ -30,3 +35,6 @@ class PacketAnalyzer:
 
     def get_dest_port(self):
         return self.layer4.dest_port
+
+    def get_content(self):
+        return self.layer7.data