+++ sniffer/network/engine.py
@@ -1,22 +1,38 @@
 import socket
+import struct
+import netifaces
 from pwn import hexdump
 from ..utils.constants import ETH_P_IP
 from .analyzer import PacketAnalyzer
-from ..exceptions.network import SLLUnsupportedError
+from ..exceptions.network import SLLUnsupportedError,\
+    UninterestingPacketException, UnsupportedVersionException
 
 
 class SnifferEngine:
-    def __init__(self):
-        self.socket = socket.socket(socket.AF_PACKET,
+    NET_INTERFACE_ANY = 'any'
+
+    def __init__(self, interface: str):
+        self.interface = interface
+        self.socket = socket.socket(socket.AF_INET,
                                     socket.SOCK_RAW,
-                                    socket.htons(ETH_P_IP))
+                                    socket.IPPROTO_TCP)
+                                    # socket.htons(ETH_P_IP))
+
+        if self.interface != SnifferEngine.NET_INTERFACE_ANY:
+            # Attach to network interface
+            self.socket.bind((
+                netifaces.ifaddresses(interface)[netifaces.AF_INET][0]['addr'],
+                0
+            ))
+    
+        # self.socket.setsockopt(socket.IPPROTO_IP, socket.IP_HDRINCL, 1)
 
     def sniff(self):
         while True:
             packet = self.socket.recvfrom(65535)[0]
-            print(hexdump(packet), len(packet))
 
             try:
+                print(hexdump(packet), len(packet), b'HTTP/' in packet)
                 analyzer = PacketAnalyzer(packet)
                 print(analyzer.get_source_mac())
                 print(analyzer.get_dest_mac())
@@ -24,5 +40,12 @@ class SnifferEngine:
                 print(analyzer.get_dest_ip())
                 print(analyzer.get_source_port())
                 print(analyzer.get_dest_port())
+                print(analyzer.get_content())
             except SLLUnsupportedError:
+                continue
+            except UnsupportedVersionException:
+                continue
+            # except struct.error:
+            #     continue
+            except UninterestingPacketException:
                 continue