+++ sniffer/utils/constants.py
@@ -9,3 +9,6 @@ PARSER_IGNORE = 1
 
 # Internet Protocol packet - Reference: /usr/include/linux/if_ether.h
 ETH_P_IP = 0x0800
+
+# Effective User ID (EUID) for root user
+ROOT_EUID = 0